requirements.txt
slack_bolt
python-jenkins
python-dotenv
